# src/web/openapi/davetiye.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Davetiye.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5052/"
    }
  ],
  "paths": {
...
    "/api/v1/public/invitations/{publicCode}": {
      "get": {
        "tags": [
          "PublicInvitationEndpoints"
        ],
        "parameters": [
          {
            "name": "publicCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "required": [
                        "status",
                        "templateKey",
                        "rendererVersion",
                        "contentSchemaVersion",
                        "content"
                      ],
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string"
                        },
                        "templateKey": {
                          "type": "string"
                        },
                        "rendererVersion": {
                          "pattern": "^-?(?:0|[1-9]\\d*)$",
                          "type": [
                            "integer",
                            "string"
                          ],
                          "format": "int32"
                        },
                        "contentSchemaVersion": {
                          "pattern": "^-?(?:0|[1-9]\\d*)$",
                          "type": [
                            "integer",
                            "string"
                          ],
                          "format": "int32"
                        },
                        "content": {
                          "required": [
                            "eventType",
                            "headline",
                            "hostNames",
                            "message",
                            "startsAt",
                            "timeZoneId",
                            "venue",
                            "programItems"
                          ],
                          "type": "object",
                          "properties": {
                            "eventType": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "headline": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "hostNames": {
                              "type": [
                                "null",
                                "array"
                              ],
                              "items": {
                                "type": "string"
                              }
                            },
                            "message": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "startsAt": {
                              "type": [
                                "null",
                                "string"
                              ],
                              "format": "date-time"
                            },
                            "timeZoneId": {
                              "type": "string"
                            },
                            "venue": {
                              "required": [
                                "name",
                                "address"
                              ],
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": [
                                    "null",
                                    "string"
                                  ]
                                },
                                "address": {
                                  "type": [
                                    "null",
                                    "string"
                                  ]
                                }
                              }
                            },
                            "programItems": {
                              "type": [
                                "null",
                                "array"
                              ],
                              "items": {
                                "required": [
                                  "title",
                                  "description",
                                  "startsAt"
                                ],
                                "type": "object",
                                "properties": {
                                  "title": {
                                    "type": [
                                      "null",
                                      "string"
                                    ]
                                  },
                                  "description": {
                                    "type": [
                                      "null",
                                      "string"
                                    ]
                                  },
                                  "startsAt": {
                                    "type": [
                                      "null",
                                      "string"
                                    ],
                                    "format": "date-time"
                                  }
                                }
                              }
                            },
                            "contacts": {
                              "type": [
                                "null",
                                "array"
                              ],
                              "items": {
                                "required": [
                                  "name",
                                  "role",
                                  "phone"
                                ],
                                "type": "object",
                                "properties": {
                                  "name": {
                                    "type": "string"
                                  },
                                  "role": {
                                    "type": [
                                      "null",
                                      "string"
                                    ]
                                  },
                                  "phone": {
                                    "type": "string"
                                  }
                                }
                              }
                            },
                            "announcement": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "faqs": {
                              "type": [
                                "null",
                                "array"
                              ],
                              "items": {
                                "required": [
                                  "question",
                                  "answer"
                                ],
                                "type": "object",
                                "properties": {
                                  "question": {
                                    "type": "string"
                                  },
                                  "answer": {
                                    "type": "string"
                                  }
                                }
                              }
                            },
                            "transportStops": {
                              "type": [
                                "null",
                                "array"
                              ],
                              "items": {
                                "required": [
                                  "name",
                                  "address",
                                  "departureTime"
                                ],
                                "type": "object",
                                "properties": {
                                  "name": {
                                    "type": "string"
                                  },
                                  "address": {
                                    "type": [
                                      "null",
                                      "string"
                                    ]
                                  },
                                  "departureTime": {
                                    "type": [
                                      "null",
                                      "string"
                                    ]
                                  }
                                }
                              }
                            }
                          }
                        },
                        "media": {
                          "type": [
                            "null",
                            "array"
                          ],
                          "items": {
                            "required": [
                              "assetId",
                              "kind",
                              "role",
                              "sortOrder"
                            ],
                            "type": "object",
                            "properties": {
                              "assetId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "kind": {
                                "type": "string"
                              },
                              "role": {
                                "type": "string"
                              },
                              "sortOrder": {
                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                "type": [
                                  "integer",
                                  "string"
                                ],
                                "format": "int32"
                              }
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "default": "unavailable"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Too Many Requests"
          }
        }
      }
    },
...
    "/api/v1/public/invitations/{publicCode}/gifts": {
      "get": {
        "tags": [
          "PublicGiftRegistryEndpoints"
        ],
        "parameters": [
          {
            "name": "publicCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicGiftRegistry"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Too Many Requests"
          }
        }
      }
    },
    "/api/v1/public/invitations/{publicCode}/gifts/reservations": {
      "post": {
        "tags": [
          "PublicGiftRegistryEndpoints"
        ],
        "parameters": [
          {
            "name": "publicCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservePublicGiftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicGiftReservationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Too Many Requests"
          }
        }
      },
      "get": {
        "tags": [
          "PublicGiftRegistryEndpoints"
        ],
        "parameters": [
          {
            "name": "publicCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicGiftGuestReservation"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Too Many Requests"
          }
        }
      }
    },
    "/api/v1/public/invitations/{publicCode}/gifts/reservations/{reservationId}": {
      "delete": {
        "tags": [
          "PublicGiftRegistryEndpoints"
        ],
        "parameters": [
          {
            "name": "publicCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "429": {
            "description": "Too Many Requests"
          }
        }
      }
    },
...
    "/api/v1/invitations": {
      "get": {
        "tags": [
          "InvitationDraftEndpoints"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDraftPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "InvitationDraftEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDraftDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/{invitationId}": {
      "get": {
        "tags": [
          "InvitationDraftEndpoints"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDraftDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/v1/invitations/{invitationId}/gifts": {
      "get": {
        "tags": [
          "CreatorGiftRegistryEndpoints"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreatorGiftItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CreatorGiftRegistryEndpoints"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGiftItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorGiftItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/invitations/{invitationId}/gifts/{itemId}": {
      "put": {
        "tags": [
          "CreatorGiftRegistryEndpoints"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGiftItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorGiftItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CreatorGiftRegistryEndpoints"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expectedRevision",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInvitationDraftRequest": {
        "required": [
          "contentSchemaVersion",
          "content",
          "templateKey"
        ],
        "type": "object",
        "properties": {
          "contentSchemaVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "content": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DraftContentInput"
              }
            ]
          },
          "templateKey": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "InvitationDraftDetails": {
        "required": [
          "id",
          "templateKey",
          "rendererVersion",
          "createdAt",
          "invitationRevision",
          "contentSchemaVersion",
          "content",
          "updatedAt",
          "contentRevision"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "rendererVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "invitationRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "contentSchemaVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "content": {
            "$ref": "#/components/schemas/DraftContentInput"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "contentRevision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "InvitationDraftPage": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationDraftSummary"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PublicGiftGuestReservation": {
        "required": [
          "reservationId",
          "itemId",
          "itemName",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "itemName": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "PublicGiftRegistry": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicGiftRegistryItem"
            }
          }
        }
      },
...
      "PublicGiftReservationResponse": {
        "required": [
          "reservationId"
        ],
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ReservePublicGiftRequest": {
        "required": [
          "itemId",
          "quantity",
          "fullName",
          "email",
          "phone"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}